# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TeamDevelop Bravo API
  description: 勤怠管理システムの REST API ドキュメント
  version: v1
  contact:
    name: TeamDevelop SRE
    email: sre@example.com
  license:
    name: Apache License Version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.teamdevelop.example.com
    description: Production server
...
paths:
...
  /api/home/overview:
    get:
      tags:
        - Home
      summary: ホーム概要
      description: ログイン中の従業員情報とお知らせ一覧を返却
      operationId: getHomeOverview
      security:
        - sessionAuth: []
      responses:
        '200':
          description: ホーム画面情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeDashboardResponse'
        '401':
          description: 認証が必要です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/home/stamps:
    post:
      tags:
        - Home
      summary: 打刻
      description: 出勤/退勤の打刻を記録
      operationId: stamp
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StampRequest'
      responses:
        '200':
          description: 打刻成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StampResponse'
        '401':
          description: 認証が必要です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/home/breaks/toggle:
    post:
      tags:
        - Home
      summary: 休憩トグル
      description: 休憩開始/終了を切り替えます
      operationId: toggleBreak
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BreakToggleRequest'
      responses:
        '204':
          description: トグル成功
        '401':
          description: 認証が必要です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 休憩開始/終了が既に登録されている場合のエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateBreakToggle:
                  summary: 重複する休憩操作
                  value:
                    message: "休憩操作は既に登録されています。画面を再読み込みして最新の勤怠状況を確認してください。"
...
components:
  schemas:
...
    EmployeeSummaryResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - admin
      properties:
        id:
          type: integer
          description: 従業員ID
        firstName:
          type: string
          description: 名
        lastName:
          type: string
          description: 姓
        email:
          type: string
          format: email
          description: メールアドレス
        admin:
          type: boolean
          description: 管理者フラグ
...
    HomeDashboardResponse:
      type: object
      required:
        - employee
        - news
      properties:
        employee:
          $ref: '#/components/schemas/EmployeeSummaryResponse'
        news:
          type: array
          items:
            $ref: '#/components/schemas/HomeNewsItem'
        attendance:
          $ref: '#/components/schemas/DailyAttendanceSnapshot'
          nullable: true
    HomeNewsItem:
      type: object
      required:
        - id
        - title
        - content
        - label
        - newsDate
        - releaseFlag
      properties:
        id:
          type: integer
          description: お知らせID
        title:
          type: string
          maxLength: 100
          description: お知らせタイトル
        content:
          type: string
          description: お知らせ内容
        label:
          type: string
          enum:
            - IMPORTANT
            - SYSTEM
            - GENERAL
          description: お知らせラベル
        newsDate:
          type: string
          format: date
          description: お知らせ日付（YYYY-MM-DD形式）
        releaseFlag:
          type: boolean
          description: 公開フラグ
    DailyAttendanceSnapshot:
      type: object
      required:
        - status
        - overtimeMinutes
      properties:
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        attendanceTime:
          type: string
          format: date-time
          nullable: true
          description: 出勤時刻（ISO 8601、JST）
        breakStartTime:
          type: string
          format: date-time
          nullable: true
          description: 休憩開始時刻（ISO 8601、JST）
        breakEndTime:
          type: string
          format: date-time
          nullable: true
          description: 休憩終了時刻（ISO 8601、JST）
        departureTime:
          type: string
          format: date-time
          nullable: true
          description: 退勤時刻（ISO 8601、JST）
        overtimeMinutes:
          type: integer
          format: int32
          description: 残業時間（分）
    AttendanceStatus:
      type: string
      description: 勤務ステータス
      enum:
        - NOT_ATTENDED
        - WORKING
        - ON_BREAK
        - FINISHED
...
    StampRequest:
      type: object
      required:
        - stampTime
        - stampType
        - nightWorkFlag
      properties:
        stampTime:
          type: string
          format: date-time
          description: 打刻時刻（ISO形式）
        stampType:
          type: string
          enum:
            - '1'
            - '2'
          description: 打刻種別（1:出勤、2:退勤）
        nightWorkFlag:
          type: string
          enum:
            - '0'
            - '1'
          description: 夜勤フラグ（0:通常、1:夜勤）
    BreakToggleRequest:
      type: object
      required:
        - timestamp
      properties:
        timestamp:
          type: string
          format: date-time
          description: 休憩トグル時刻（ISO 8601形式）
    StampResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: 打刻結果メッセージ
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: エラーメッセージ
        details:
          type: object
          additionalProperties: true
          description: エラー詳細情報
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: SESSION